Tidy ups

+++ app/routes.php
@@ -6,8 +6,8 @@ $app->group('', function () use ($app) {
 
     $controller = new CrSrc\Controller\IndexController($app);
 
-    $app->get('/', $controller('index'));
-    $app->get('/contact', $controller('contact'));
+    $app->get('/', $controller('index'))->setName('index');
+    $app->get('/contact', $controller('contact'))->setName('contact');
 });
 
 // index routes (homepage, about, etc)
@@ -15,9 +15,9 @@ $app->group('/session', function () use ($app) {
 
     $controller = new CrSrc\Controller\SessionController($app);
 
-    $app->get('/login', $controller('login'));
-    $app->post('', $controller('post'));
-    $app->delete('', $controller('delete'));
+    $app->get('/login', $controller('login'))->setName('session_login');
+    $app->post('', $controller('post'))->setName('session_post');
+    $app->delete('', $controller('delete'))->setName('session_delete');
 });
 
 // create resource method for Slim::resource($route, $name)
@@ -25,8 +25,8 @@ $app->group('/articles', function () use ($app) {
 
     $controller = new CrSrc\Controller\ArticlesController($app);
 
-    $app->get('/{id:[0-9]+}', $controller('show'));
-    $app->get('/{id:[0-9]+}/{slug}', $controller('show'));
+    $app->get('/{id:[0-9]+}', $controller('show'))->setName('articles_show');
+    $app->get('/{id:[0-9]+}/{slug}', $controller('show'))->setName('articles_show_wslug');
 });
 
 // users routes (eg. register)
@@ -34,25 +34,26 @@ $app->group('/users', function () use ($app) {
 
     $controller = new CrSrc\Controller\UsersController($app);
 
-    $app->get('/create', $controller('create'));
-    $app->post('', $controller('post'));
+    $app->get('/create', $controller('create'))->setName('users_create');
+    $app->post('', $controller('post'))->setName('users_post');
 });
 
-// admin routes
+// admin routes -- invokes auth middleware
 $app->group('/admin', function () use ($app) {
 
     // admin/articles routes
     $app->group('/articles', function () use ($app) {
 
         $controller = new CrSrc\Controller\Admin\ArticlesController($app);
+        $container = $app->getContainer();
 
-        $app->get('', $controller('index'));
-        $app->get('/{id:[0-9]+}', $controller('show'));
-        $app->get('/create', $controller('create'));
-        $app->get('/{id:[0-9]+}/edit', $controller('edit'));
+        $app->get('', $controller('index'))->setName('admin_articles_index');
+        $app->get('/{id:[0-9]+}', $controller('show'))->setName('admin_articles_show');
+        $app->get('/create', $controller('create'))->setName('admin_articles_create');
+        $app->get('/{id:[0-9]+}/edit', $controller('edit'))->setName('admin_articles_edit');
 
-        $app->post('', $controller('post'));
-        $app->put('/{id:[0-9]+}', $controller('update'));
-        $app->delete('/{id:[0-9]+}', $controller('delete'));
+        $app->post('', $controller('post'))->setName('admin_articles_post');
+        $app->put('/{id:[0-9]+}', $controller('update'))->setName('admin_articles_update');
+        $app->delete('/{id:[0-9]+}', $controller('delete'))->setName('admin_articles_delete');
     });
 })->add( new \CrSrc\Middleware\Auth() );
